tests/requirements.txt
pytest==4.4.1
python-consul==1.1.0
pyzmq==18.0.1
